13-10 push code

+++ Source/EmailCampaign.WebApplication/Controllers/PermissionController.cs
@@ -5,18 +5,22 @@ using EmailCampaign.Domain.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 using AuthorizeAttribute = Microsoft.AspNetCore.Authorization.AuthorizeAttribute;
 using EmailCampaign.Infrastructure.Data.Repositories;
+using EmailCampaign.Infrastructure.Data.Services.LogsService;
+using Microsoft.AspNetCore.Mvc.Filters;
 
 namespace EmailCampaign.WebApplication.Controllers
 {
-    //[Microsoft.AspNetCore.Authorization.Authorize]
+    [Microsoft.AspNetCore.Authorization.Authorize]
     public class PermissionController : Controller
     {
         private readonly IPermissionRepository _permissionRepository;
+        private readonly IRolePermissionRepository _rolePermissionRepository;
         private readonly IMapper _mapper;
-        public PermissionController(IPermissionRepository permissionRepository, IMapper mapper)
+        public PermissionController(IPermissionRepository permissionRepository, IMapper mapper, IRolePermissionRepository rolePermissionRepository)
         {
             _permissionRepository = permissionRepository;
             _mapper = mapper;
+            _rolePermissionRepository = rolePermissionRepository;
         }
 
         [Authorize("ViewPermission")]
@@ -26,7 +30,7 @@ namespace EmailCampaign.WebApplication.Controllers
             return View(roleList);
         }
 
-        [Authorize("AddEditPermission")]
+        //[Authorize("AddEditPermission")]
         public async Task<IActionResult> GetPermissionById(Guid id)
         {
             Permission permission = await _permissionRepository.GetPermissionAsync(id);
@@ -54,10 +58,21 @@ namespace EmailCampaign.WebApplication.Controllers
 
             if (permission == null)
             {
-                TempData["Message"] = "";
+                TempData["ErrorMessage"] = "Permission create process failed !!";
                 return View("AddPermission");
             }
 
+            var activityLogAttribute = new ActivityLogAttribute("Create", "New Permission created with name {0} Created by {1}", permission.ControllerName +  " (" + permission.Id + ")");
+            var context = new ActionExecutingContext(
+            new ActionContext(HttpContext, RouteData, new Microsoft.AspNetCore.Mvc.Abstractions.ActionDescriptor()),
+            new List<IFilterMetadata>(),
+            new Dictionary<string, object> { },
+            this
+            );
+            await activityLogAttribute.LogActivityAsync(context);
+
+
+            TempData["SuccessMessage"] = "Permission successfully created.";
             return RedirectToAction("Index");
         }
 
@@ -76,13 +91,32 @@ namespace EmailCampaign.WebApplication.Controllers
         [Authorize("DeletePermission")]
         public async Task<IActionResult> Delete(Guid id)
         {
-            var isDeleted = await _permissionRepository.DeletePermissionAsync(id);
+            RolePermission rolePermission = await _rolePermissionRepository.GetItemByPermissionId(id);
 
-            if (!isDeleted)
+            if(rolePermission != null)
             {
-                TempData["Message"] = "";
+                TempData["ErrorMessage"] = "Permission already mapped with RolePermission. so can not be delete.";
+                return RedirectToAction("Index");
+            }
+
+            var permission = await _permissionRepository.DeletePermissionAsync(id);
+
+            if (permission == null)
+            {
+                TempData["ErrorMessage"] = "Permission delete process failed.";
                 return View("Index");
             }
+
+            var activityLogAttribute = new ActivityLogAttribute("Delete", "Permission detalis Updated with name {0} Created by {1}", permission.ControllerName + " (" + permission.Id + ")");
+            var context = new ActionExecutingContext(
+            new ActionContext(HttpContext, RouteData, new Microsoft.AspNetCore.Mvc.Abstractions.ActionDescriptor()),
+            new List<IFilterMetadata>(),
+            new Dictionary<string, object> { },
+            this
+            );
+            await activityLogAttribute.LogActivityAsync(context);
+
+            TempData["SuccessMessage"] = "Permission deleted successfully.";
             return RedirectToAction("Index");
         }
